🔭 Add initial previewer
It simply runs `gt log` for every branch you select. I didn't find a
`gt` option to restrict logging to the current branch.

An alternative could be `git log`, but that's not very clean when using
graphite. It includes `graphite-base` branches and so on.

But maybe it's enough for a first version.